backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
starlette==0.27.0

# Database
pymongo==4.5.0
dnspython==2.8.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.8

# Environment & HTTP
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7

# Utilities
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
typing-extensions==4.8.0
anyio==3.7.1

# Development
h11==0.14.0
sniffio==1.3.0